Refactor project into segments, fix CI, and reorganize directories.
