+++ src/web/components/Attribute/segments/AttributeLabel.jsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+import { text } from '../../../theme/helpers';
+
+const AttributeLabel = styled.label`
+  font-weight: 800;
+  color: ${text('normal')};
+`;
+
+export default AttributeLabel;
