+++ src/web/components/Attribute/segments/AttributeRow.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const AttributeRow = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+export default AttributeRow;
