+++ src/web/components/Attribute/segments/index.js
@@ -1,0 +1,2 @@
+export { default as AttributeRow } from './AttributeRow';
+export { default as AttributeLabel } from './AttributeLabel';
