+++ src/web/components/Dot/segments/index.js
@@ -1,0 +1,1 @@
+export { default as DotContainer } from './DotContainer';
