+++ src/web/components/DotGroup/segments/DotGroupContainer.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const DotGroupContainer = styled.div`
+  display: flex;
+  width: min-content;
+`;
+
+export default DotGroupContainer;
