+++ src/web/components/DotGroup/segments/index.js
@@ -1,0 +1,1 @@
+export { default as DotGroupContainer } from './DotGroupContainer';
