+++ src/web/components/Layout/segements/Title.jsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+import { bgColor } from '../../../theme';
+
+const Title = styled.header`
+  grid-area: title;
+
+  color: white;
+
+  background-color: ${bgColor('title')};
+
+  z-index: 1;
+
+  & > h1 {
+    font-size: 1.5rem;
+  }
+`;
+
+export default Title;
