+++ src/web/components/Layout/segements/index.js
@@ -1,0 +1,2 @@
+export { default as LayoutGrid } from './LayoutGrid';
+export { default as Title } from './Title';
