+++ src/web/components/NavButton/segments/index.js
@@ -1,0 +1,1 @@
+export { default as NavButtonBase } from './NavButtonBase';
