+++ src/web/components/NotchedBox/segments/NotchedBoxBorder.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const NotchedBoxContainer = styled.div`
+  position: relative;
+  padding: 2px;
+  margin-bottom: 25px;
+`;
+
+export default NotchedBoxContainer;
