+++ src/web/components/NotchedBox/segments/index.js
@@ -1,0 +1,3 @@
+import { default as NotchedBoxBorder } from './NotchedBoxBorder';
+import { default as NotchedBoxContainer } from './NotchedBoxContainer';
+import { default as NotchedBoxContent } from './NotchedBoxContent';
