+++ src/web/components/Panel/segments/PanelDiv.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const PanelDiv = styled.div`
+  background-color: ${({ theme }) => theme.colors.background.panel};
+`;
+
+export default PanelDiv;
