+++ src/web/components/Panel/segments/index.js
@@ -1,0 +1,1 @@
+export { default as PanelDiv } from './PanelDiv';
