+++ src/web/components/PinnableButton/PinnableButton.jsx
@@ -1,162 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Icon } from '@chakra-ui/core';
-import styled from 'styled-components';
 import { useTheme } from '@chakra-ui/core';
 
-import { bgColor, accent, text } from '../../theme/helpers';
-
-const PinnableButtonContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-
-  width: min-content;
-  font-size: 0px;
-  white-space: nowrap;
-`;
-
-const PinnableButtonLeft = styled.button.attrs({
-  className: 'PinnableButtonPin',
-})`
-  height: 1.5rem;
-  width: 0.75rem;
-
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-
-  word-break: keep-all;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-
-  padding: 0.25rem 0;
-  margin: 0;
-
-  background-color: ${bgColor('primary')};
-  color: ${text('normal')};
-
-  border: none;
-  box-shadow: -0.06rem 0 0 0 ${accent('primary')};
-  border-radius: 4rem 0 0 4rem;
-
-  &:focus {
-    outline: 0;
-  }
-
-  & > svg {
-    margin-right: 0.1rem;
-  }
-
-  .pinned > & {
-    box-shadow: -0.06rem 0 0 0.12rem ${accent('primary')};
-
-    & > svg {
-      color: ${accent('primary')};
-    }
-  }
-`;
-
-const PinnableButtonRight = styled.button.attrs({
-  className: 'PinnableButtonRoute',
-})`
-  height: 1.5rem;
-  width: 0.75rem;
-
-  width: min-content;
-  max-width: 6rem;
-  min-width: 2rem;
-
-  word-break: keep-all;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-
-  padding: 0.25rem 0.2rem;
-  margin: 0;
-
-  color: ${text('normal')};
-  text-shadow: black 0.05rem 0.05rem 0.05rem;
-  background-color: ${bgColor('button')};
-  border-radius: 0 4rem 4rem 0;
-  border: none;
-
-  background-size: 200% 100%;
-  background-image: linear-gradient(
-    to left,
-    ${bgColor('primary')} 50%,
-    ${bgColor('title')} 50%
-  );
-
-  transition: background-position 0.2s;
-
-  &:focus {
-    outline: 0;
-  }
-
-  .PinnableButton:hover > & {
-    background-position: -100% 0;
-  }
-
-  .pinned > & {
-    box-shadow: 0.03rem -0.12rem 0 0 ${accent('primary')},
-      0.03rem 0.12rem 0 0 ${accent('primary')},
-      0.07rem 0 0 0.07rem ${accent('primary')};
-  }
-
-  .active > & {
-    color: ${accent('primary')};
-    text-decoration: underline;
-    font-weight: bold;
-  }
-`;
-
-const CloseButton = styled.button.attrs({
-  className: 'PinnableButtonClose',
-})`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  height: 0.6rem;
-  width: 1.2rem;
-
-  position: absolute;
-  right: 0.7rem;
-  top: -0.6rem;
-
-  font-size: 0.6rem;
-  color: ${text('normal')};
-  text-shadow: black 0.05rem 0.05rem 0.02rem;
-  background-color: ${bgColor('button')};
-
-  border: none;
-  border-radius: 1.2rem 1.2rem 0 0;
-
-  &:focus {
-    outline: 0;
-  }
-
-  .pinned > & {
-    background-color: ${accent('primary')};
-    background-image: none;
-  }
-
-  background-size: 200% 100%;
-  background-image: linear-gradient(
-    to left,
-    ${bgColor('primary')} 50%,
-    ${bgColor('title')} 50%
-  );
-
-  transition: background-position 0.08s;
-
-  .PinnableButton:hover > & {
-    background-position: -100% 0;
-    transition: background-position 0.12s 0.08s;
-  }
-`;
+import {
+  CloseButton,
+  PinnableButtonContainer,
+  PinnableButtonLeft,
+  PinnableButtonRight,
+} from './segments';
 
 /**
  * PinnableButton can be used to allow the ST or player to get to
