+++ src/web/components/PinnableButton/segments/PinnableButtonContainer.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const PinnableButtonContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+
+  width: min-content;
+  font-size: 0px;
+  white-space: nowrap;
+`;
+
+export default PinnableButtonContainer;
