+++ src/web/components/PinnableButton/segments/index.js
@@ -1,0 +1,4 @@
+export { default as CloseButton } from './CloseButton';
+export { default as PinnableButtonContainer } from './PinnableButtonContainer';
+export { default as PinnableButtonLeft } from './PinnableButtonLeft';
+export { default as PinnableButtonRight } from './PinnableButtonRight';
