--- src/web/components/SecondEditionCombat/index.js
@@ -1,1 +1,0 @@
-export { default } from './SecondEditionCombat';
