+++ src/web/components/Sidebar/segments/NavSettings.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const NavSettings = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  justify-content: flex-end;
+  align-items: center;
+  width: 55px;
+`;
+
+export default NavSettings;
