+++ src/web/components/Sidebar/segments/index.js
@@ -1,0 +1,6 @@
+export { default as ExpandableNav } from './ExpandableNav';
+export { default as ExpandArrow } from './ExpandArrow';
+export { default as Nav } from './Nav';
+export { default as NavButtons } from './NavButtons';
+export { default as NavLinkStyled } from './NavLinkStyled';
+export { default as NavSettings } from './NavSettings';
