+++ src/web/components/VerticalDivider/segments/index.js
@@ -1,0 +1,1 @@
+export { default as VerticalDividerElement } from './VerticalDividerElement';
