+++ src/web/components/routes/CharactersRoute/index.js
